+++ jgclark.Dashboard/src/dashboardTriggers.js
@@ -1,16 +1,57 @@
 // @flow
 //-----------------------------------------------------------------------------
 // Dashboard triggering
-// Last updated 1.3.2023 for v0.3.x by @jgclark
+// Last updated 24.3.2023 for v0.3.x by @jgclark
 //-----------------------------------------------------------------------------
 
 import pluginJson from '../plugin.json'
 import { showDashboardHTML } from './dashboardHTML'
-import { clo, JSP, logDebug, logError, logInfo, logWarn } from '@helpers/dev'
+import { clo, JSP, /*logDebug,*/ logError, logInfo, logWarn } from '@helpers/dev'
 import { rangeToString } from '@helpers/general'
 import { selectedLinesIndex } from '@helpers/NPparagraph'
 import { isHTMLWindowOpen } from '@helpers/NPWindows'
-import { RE_ANY_TYPE_OF_TASK_OR_CHECKLIST_MARKER } from '@helpers/regex'
+import {
+  RE_ANY_TYPE_OF_OPEN_TASK_OR_CHECKLIST_MARKER_MULTI_LINE,
+  RE_ANY_TYPE_OF_CLOSED_TASK_OR_CHECKLIST_MARKER_MULTI_LINE
+} from '@helpers/regex'
+import plugin from "@babel/core/lib/config/plugin";
+
+/**
+ * Local version of log, turned on by a more specific setting
+ * @param {any} pluginJson 
+ * @param {string} message 
+ */
+function logDebug(pluginJson: any, message: string): void {
+  if (pluginJson?.settings?.triggerLogLevel) {
+    console.log(message)
+    // } else {
+    //   clo(pluginJson)
+    //   console.log(message)
+  }
+}
+
+/**
+ * Return true if some task/checklist items have been added or completed when comparing 'previousContent' to 'currentContent'.
+ * @param {string} previousContent 
+ * @param {string} currentContent 
+ * @returns {boolean} changed?
+ */
+function changeToNumberOfOpenItems(previousContent: string, currentContent: string): boolean {
+  const prevOpenNum = numberOfOpenItems(previousContent)
+  const currentOpenNum = numberOfOpenItems(currentContent)
+  // logDebug(pluginJson, `prevOpenNum: ${prevOpenNum} / currentOpenNum: ${currentOpenNum} ->  ${String(prevOpenNum - currentOpenNum)}`)
+  return (prevOpenNum != currentOpenNum)
+}
+
+/**
+ * Return number of open items in a multi-line string
+ * @param {number} content 
+ * @returns {number}
+ */
+function numberOfOpenItems(content: string): number {
+  const res = Array.from(content.matchAll(RE_ANY_TYPE_OF_OPEN_TASK_OR_CHECKLIST_MARKER_MULTI_LINE))
+  return res ? res.length : 0
+}
 
 /**
  * Decide whether to update Dashboard, to be called by an onSave or onChange trigger
@@ -50,23 +91,25 @@ export function decideWhetherToUpdateDashboard(): void {
       return
     }
 
-    // Get changed ranges
-    const ranges = NotePlan.stringDiff(previousContent, latestContent)
-    if (!ranges || ranges.length === 0) {
-      logDebug(`No ranges returned, so stopping.`)
-      return
-    }
-    const earliestStart = ranges[0].start
-    let latestEnd = ranges[ranges.length - 1].end
-    const overallRange: TRange = Range.create(earliestStart, latestEnd)
-    logDebug('dashboard/decideWhetherToUpdateDashboard', `- overall changed content from ${rangeToString(overallRange)}`)
+    // Decide if there are relevant changes
 
-    // earlier method for changedExtent based on character region, which didn't seem to always include all the changed parts.
-    let changedExtent = latestContent?.slice(earliestStart, latestEnd)
+    // for v1 + v2: Get changed ranges
+    // const ranges = NotePlan.stringDiff(previousContent, latestContent)
+    // if (!ranges || ranges.length === 0) {
+    //   logDebug(pluginJson, `No ranges returned, so stopping.`)
+    //   return
+    // }
+    // const earliestStart = ranges[0].start
+    // let latestEnd = ranges[ranges.length - 1].end
+    // const overallRange: TRange = Range.create(earliestStart, latestEnd)
+    // logDebug('dashboard/decideWhetherToUpdateDashboard', `- overall changed content from ${rangeToString(overallRange)}`)
+
+    // v1: for changedExtent based on character region, which didn't seem to always include all the changed parts.
+    // let changedExtent = latestContent?.slice(earliestStart, latestEnd)
     // Editor.highlightByIndex(earliestStart, latestEnd - earliestStart)
-    logDebug('dashboard/decideWhetherToUpdateDashboard', `Changed content  (method 1): <${changedExtent}>`)
+    // logDebug('dashboard/decideWhetherToUpdateDashboard', `Changed content  (method 1): <${changedExtent}>`)
 
-    // // Newer method uses changed paragraphs: this will include more than necessary, but that's more useful in this case
+    // v2: Newer method uses changed paragraphs: this will include more than necessary, but that's more useful in this case
     // changedExtent = ''
     // const [startParaIndex, endParaIndex] = selectedLinesIndex(overallRange, Editor.paragraphs)
     // logDebug('dashboard/decideWhetherToUpdateDashboard', `- changed lines ${startParaIndex}-${endParaIndex}`)
@@ -76,12 +119,17 @@ export function decideWhetherToUpdateDashboard(): void {
     // }
     // logDebug('dashboard/decideWhetherToUpdateDashboard', `Changed content (method 2): <${changedExtent}>`)
 
-    // Is the change to the note in question one that adds or completes or removes a task or checklist?
-    const isThisChangeSignificant = RE_ANY_TYPE_OF_TASK_OR_CHECKLIST_MARKER.test(changedExtent) // TODO:
+    // v3: Doesn't use ranges. This compares the whole of the current and previous content, asking are there a different number of open items?
+    // (This avoids firing when simply moving task/checklist items around.)
+    const isThisChangeSignificant = changeToNumberOfOpenItems(previousContent, latestContent)
+
     if (isThisChangeSignificant) {
       // Update the dashboard
-      logDebug(pluginJson, `Have decided to update dashboard.`)
+      logDebug(pluginJson, `WILL update dashboard.`)
       showDashboardHTML()
+    }
+    else {
+      logDebug(pluginJson, `Won't update dashboard.`)
     }
   }
   catch (error) {