+++ jgclark.Dashboard/src/index.js
@@ -3,13 +3,14 @@
 // ----------------------------------------------------------------------------
 // Dashboard plugin for NotePlan
 // Jonathan Clark
-// last updated 15.3.2023 for v0.3.1, @jgclark
+// last updated 22.3.2023 for v0.3.2, @jgclark
 // ----------------------------------------------------------------------------
 
 // allow changes in plugin.json to trigger recompilation
 import pluginJson from '../plugin.json'
 import { JSP, logDebug, logError, logInfo, logWarn } from '@helpers/dev'
 import { getPluginJson, pluginUpdated, updateSettingData } from '@helpers/NPConfiguration'
+import { editSettings } from '@helpers/NPSettings'
 import { showMessage } from '@helpers/userInput'
 
 export { logWindowsList } from '@helpers/NPWindows'
@@ -41,19 +42,27 @@ export function onSettingsUpdated(): void {
   // Placeholder only to stop error in logs
 }
 
-// export async function onUpdateOrInstall(): Promise<void> {
-//   try {
-//     logDebug(pluginJson, `${thisPluginID}: onUpdateOrInstall running`)
-//     // Try updating settings data
-//     const updateSettings = updateSettingData(pluginJson)
-//     logDebug(pluginJson, `${thisPluginID}: onUpdateOrInstall updateSettingData code: ${updateSettings}`)
+export async function onUpdateOrInstall(): Promise<void> {
+  try {
+    // Tell user the plugin has been updated
+    if (pluginJson['plugin.lastUpdateInfo'] !== 'undefined') {
+      await showMessage(pluginJson['plugin.lastUpdateInfo'], 'OK, thanks', `Plugin ${pluginJson['plugin.name']}\nupdated to v${pluginJson['plugin.version']}`)
+    }
+  } catch (error) {
+    logError(pluginJson, JSP(error))
+  }
+}
 
-//     // Tell user the plugin has been updated
-//     if (pluginJson['plugin.lastUpdateInfo'] !== 'undefined') {
-//       await showMessage(pluginJson['plugin.lastUpdateInfo'], 'OK, thanks', `Plugin ${pluginJson['plugin.name']}\nupdated to v${pluginJson['plugin.version']}`)
-//     }
-//   } catch (error) {
-//     logError(pluginJson, error)
-//   }
-//   logDebug(pluginJson, `${thisPluginID}: onUpdateOrInstall finished`)
-// }
+/**
+ * Update Settings/Preferences (for iOS etc)
+ * Plugin entrypoint for command: "/<plugin>: Update Plugin Settings/Preferences"
+ * @author @dwertheimer
+ */
+export async function updateSettings() {
+  try {
+    logDebug(pluginJson, `updateSettings running`)
+    await editSettings(pluginJson)
+  } catch (error) {
+    logError(pluginJson, JSP(error))
+  }
+}